Move code into repo's root
