--- data/search.csv
@@ -1,39 +1,0 @@
-# label:domain:query-keys
-AOL:aol,aolsvc:query,encquery,q
-About:about:terms
-Alice:alice:qs
-AllTheWeb:alltheweb:q
-Altavista:altavista:q
-Ask:ask:q
-Baidu:baidu:wd
-Bing:bing:q
-CNN:cnn:query
-Daum:daum:q
-DuckDuckGo:duckduckgo:q,*
-Ekolay:ekolay:q
-Eniro:eniro:search_word
-Google:google:q,*
-Kvasir:kvasir:q
-#Live:bing:q # duplicate of Bing:bing:q
-Lycos:lycos:query
-MSN:msn:q
-Mamma:mamma:query
-Mynet:mynet:q
-Najdi:najdi:q
-Naver:naver:query
-Netscape:netscape:query,q
-O*NET:onetcenter:qt
-Ozu:ozu:q
-PCHome:pchome:q
-Rambler:rambler:words
-Search:search:q
-Sesam:sesam:q
-Seznam:seznam:q
-Szukacz:szukacz:q
-Terra:terra:query
-Virgilio:virgilio:qs
-Voila:voila:rdata
-Wirtulana Polska:wp:szukaj
-Yahoo!:yahoo:p,*
-Yam:yam:k
-Yandex:yandex:text
